azure-functions-deployment/functions/requirements.txt
# Azure Functions requirements
azure-functions==1.17.0

# FastAPI and dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis
redis==5.0.1

# Azure Storage
azure-storage-blob==12.19.0
azure-identity==1.15.0

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
httpx==0.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# AI/Content
openai==1.3.7
beautifulsoup4==4.12.2
lxml==4.9.3
